Bring back parsing of OneLifeData7 data, so we can use it in our filters
There's still work to do. For instance, we may want to try to combine
all the object data into a single struct that has onelifedata7 and
twotech data as substructs. Not sure if we get much from that though.
We'd had to have them be Options in case one of them parses something or
has an entry the other doesn't. We'd also want to make sure we have the
entire initial data set for both sets of data saved in variables that
don't get shadowed away.

This extra deserialization also adds 15 seconds to the program runtime,
even without further filtering. It would be REALLY great to add the
ability to specify pre-parsed Rust object vector data files. It would
save a lot of time per-run.